Migrate Code to New File
Additionally, cleaned up some code.
